% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/analyze-data.R
\name{analyze_data}
\alias{analyze_data}
\title{Analyze data}
\usage{
analyze_data(data)
}
\arguments{
\item{data}{A data frame containing type, asset, date and return columns.}
}
\value{
TODO
}
\description{
TODO
}
