% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/combine-data.R
\name{combine_data}
\alias{combine_data}
\title{Combine data}
\usage{
combine_data(...)
}
\arguments{
\item{...}{The downloaded data as data frames.}
}
\value{
A data frame containing the following columns:
\itemize{
\item asset
\item date
\item return
}

and arranged by asset and date.
}
\description{
Bind downloaded data and arrange by asset and date.
}
