% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/complement-data.R
\name{complement_data}
\alias{complement_data}
\title{Complement data}
\usage{
complement_data(data, config)
}
\arguments{
\item{data}{A data frame containing asset, date and return columns.}

\item{config}{A list of lists containing the complement methods and arguments. They are
\code{method} (method name, one of \code{proxy}, \code{spread}, \code{regress} or \code{mix}) and
\code{arguments} (method arguments).}
}
\value{
A data frame containing the following columns:
\itemize{
\item asset
\item date
\item return
}

and arranged by asset and date.
}
\description{
Complement the data with the following functionalities:
\itemize{
\item Proxy: Add "proxied" asset that is a copy of "base" asset
\item Spread: Spread non-daily returns to daily returns of "spreaded" asset
\item Regress: Extend "regressed" asset returns as a linear function of "base"
asset returns
\item Mix: Add "mixed" asset whose returns are the average of "base" assets
returns for each day.
}
}
