% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-cumulative-dca.R
\name{cumulative_dca}
\alias{cumulative_dca}
\title{Cumulative DCA}
\usage{
cumulative_dca(xs)
}
\arguments{
\item{xs}{The sequence of values.}
}
\value{
The cumulative DCA of the sequence X.
}
\description{
Compute the cumulative DCA of a sequence X.
}
\details{
The cumulative DCA is described by the follwing equations.

DCA[1] = X[1]

DCA[2] = X[1] • X[2] + X[2]

DCA[3] = X[1] • X[2] • X[3] + X[2] • X[3] + X[3]

DCA[N] = X[1] • ... • X[N] + X[2] • ... • X[N] + ... + X[N-1] • X[N] + X[N]

DCA[n] = sum(i = 1 to n) prod(j = i to n) X[j]

DCA[n] = DCA[n-1] * X[n] + X[n]
}
