% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/download-b3-data.R
\name{download_b3_data}
\alias{download_b3_data}
\title{Download data from the Brazil Stock Exchange (B3)}
\usage{
download_b3_data(config)
}
\arguments{
\item{config}{A list of lists containing the B3 data arguments. They are \code{asset} (asset
name), \code{index} (index ID, as per B3 site) and \code{first_year} (first year with
available data, as per B3 site).}
}
\value{
A data frame containing the following columns:
\itemize{
\item asset
\item date
\item return
}

and arranged by asset and date.
}
\description{
Download index data from the Brazil Stock Exchange (B3).
}
\details{
This function downloads the data directly from an internal API of B3 that
feeds some tables in the B3 site (e.g. \href{http://www.b3.com.br/pt_br/market-data-e-indices/indices/indices-amplos/indice-ibovespa-ibovespa-estatisticas-historicas.html}{IBOV data table}).
}
