% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/download-bcb-data.R
\name{download_bcb_data}
\alias{download_bcb_data}
\title{Download data from the Central Bank of Brazil (BCB)}
\usage{
download_bcb_data(config)
}
\arguments{
\item{config}{A list of lists containing the BCB data arguments. They are \code{asset} (asset
name), \code{series} (series number, as per sgs api) and \code{unit} (asset unit,
either return or price).}
}
\value{
A data frame containing the following columns:
\itemize{
\item asset
\item date
\item return
}

and arranged by asset and date.
}
\description{
Download macro and index data from the Central Bank of Brazil (BCB).
}
\details{
This function downloads the data directly from the \href{https://dadosabertos.bcb.gov.br/dataset}{sgs api}. You can check the available series
manually at \href{https://www3.bcb.gov.br/sgspub/}{sgs site}.
}
