% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pack-data.R
\name{pack_data}
\alias{pack_data}
\title{Pack data}
\usage{
pack_data(data, benchmark, assets, simulate_portfolio = FALSE)
}
\arguments{
\item{data}{A data frame containing asset, date and return columns.}

\item{benchmark}{Benchmark name.}

\item{assets}{Asset names.}

\item{simulate_portfolio}{TRUE if should include a risk parity portfolio into the packed data.}
}
\value{
A data frame containing the following columns:
\itemize{
\item type
\item asset
\item date
\item return
}

and arranged by type, asset and date.
}
\description{
TODO
}
