% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RiskParityBrazil-package.R
\docType{package}
\name{RiskParityBrazil-package}
\alias{RiskParityBrazil}
\alias{RiskParityBrazil-package}
\title{RiskParityBrazil: An analysis of risk parity applied to the Brazilian stock market}
\description{
A complete pipeline that downloads financial data, simulates
    the historical performance of risk parity portfolios in the Brazilian
    stock market and presents the results in monthly updated dashboards
    along with a web article. Implemented as a R package on top of the R
    package targets (pipeline management). Deployed with GitHub Pages.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/GiuseppeTT/RiskParityBrazil}
  \item \url{https://giuseppett.github.io/RiskParityBrazil/}
}

}
\author{
\strong{Maintainer}: Giuseppe Tomio \email{g.tinti.tomio@gmail.com}

}
\keyword{internal}
