% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-trading-days-per.R
\docType{data}
\name{trading_days_per}
\alias{trading_days_per}
\title{Trading day counts}
\format{
An object of class \code{list} of length 3.
}
\usage{
trading_days_per
}
\description{
TODO
}
\keyword{datasets}
